+++ packages/switch/src/index.ts
@@ -1,0 +1,4 @@
+import { Switch } from "./switch.js";
+
+export { Switch };
+export type { SwitchDetail } from "./types.js";